+++ src/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { App } from './app'
+import { App } from './app1'
 
 ReactDOM.render(<App />, document.getElementById('root'))
